ADD:メビウス呼び名設定機能・レベルアップ時報酬(下準備)

+++ src/com/github/unchama/seichiassist/util/ExperienceManager.java
@@ -184,7 +184,7 @@ public class ExperienceManager {
 		Player player = getPlayer();
 
 		int lvl = player.getLevel();
-		int cur = getXpForLevel(lvl) + (int) Math.round(getXpNeededToLevelUp(lvl) * player.getExp());
+		int cur = getXpForLevel(lvl) + Math.round(getXpNeededToLevelUp(lvl) * player.getExp());
 		return cur;
 	}
